+++ components/02-elements/dropdown-list/dropdown-list.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   status: 'wip',
   context: {
+    script: true,
     dropdowns: [
       {
         itemTag: 'a',