+++ components/04-views/catalog/grid/grid.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  name: 'catalog-grid',
+  name: 'grid-view',
   context: {
     product: {
       tag: 'li',